# docs/api.yaml
openapi: 3.0.0
info:
  title: Pollution API
  version: 1.0.0
paths:
  /v1/pollution:
    get:
      summary: Gets pollution data for the nearest city to a given set of coordinates
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: string
        - name: longitude
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A PollutionData object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollutionData"
        "400":
          description: Bad Request
        "500":
          description: An error occurred while retrieving the data

  /v1/pollution/paris/max:
    get:
      summary: Get the time when the pollution was at its highest level in Paris
      responses:
        "200":
          description: A JSON object containing the date and time of the highest pollution level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollutionMaxResponse"
        "404":
          description: No pollution data found for Paris
        "500":
          description: An error occurred while retrieving the data
components:
  schemas:
    PollutionData:
      type: object
      properties:
        Result:
          type: object
          properties:
            Pollution:
              type: object
              properties:
                ts:
                  type: string
                  example: "2024-02-02T15:33:19Z"
                aqius:
                  type: number
                  example: 42
                mainus:
                  type: string
                  example: "p2"
                aqicn:
                  type: number
                  example: 20
                maincn:
                  type: string
                  example: "pm25"
    PollutionMaxResponse:
      type: object
      properties:
        Result:
          type: object
          properties:
            Datetime:
              type: string
              format: date-time
              example: "2024-02-02T15:33:19Z"